Mise a jour de 6S en double (option -r8 dans f2c)

+++ Utilities/otb6S/6SV1.0B/DISCRE.c
@@ -6,7 +6,6 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-
 /* OTB patches: replace "f2c.h" by "otb_6S.h" */
 /*#include "f2c.h"*/
 #include "otb_6S.h"
@@ -32,7 +31,7 @@ extern "C" {
 /*<       double precision xx >*/
 /*<       integer it,nt >*/
 /*<       if (ha.ge.7.) then >*/
-    if (*ha >= (float)7.) {
+    if (*ha >= 7.) {
 /*<    >*/
 	print_error__("check aerosol measurements or plane altitude", 44L);
 /*<           return >*/
@@ -42,16 +41,16 @@ extern "C" {
 /*<       if (it.eq.0) then >*/
     if (*it == 0) {
 /*<          dt=1.e-17 >*/
-	dt = (float)1e-17;
+	dt = 1e-17;
 /*<          else >*/
     } else {
 /*<          dt=2.*(ta+tr-yy)/(nt-it+1.) >*/
-	dt = (*ta + *tr - *yy) * (float)2. / (*nt - *it + (float)1.);
+	dt = (*ta + *tr - *yy) * 2. / (*nt - *it + 1.);
 /*<       endif >*/
     }
 /*<   99  dt=dt/2. >*/
 L99:
-    dt /= (float)2.;
+    dt /= 2.;
 /*<       ti=yy+dt >*/
     ti = *yy + dt;
 /*<       y1=ppp2 >*/
@@ -60,7 +59,7 @@ L99:
     y3 = *ppp1;
 /*<   706 y2=(y1+y3)*0.5 >*/
 L706:
-    y2 = (y1 + y3) * (float).5;
+    y2 = (y1 + y3) * .5;
 /*<       xx=-y2/ha >*/
     xx = -y2 / *ha;
 /*<       if (xx.lt.-18) then >*/
@@ -76,7 +75,7 @@ L706:
 /*<       xd=abs(ti-x2) >*/
     xd = (d__1 = ti - x2, abs(d__1));
 /*<       if(xd.lt.0.00001) go to 705 >*/
-    if (xd < (float)1e-5) {
+    if (xd < 1e-5) {
 	goto L705;
     }
 /*<       if(ti-x2) 701,703,703 >*/
@@ -99,8 +98,8 @@ L703:
 L705:
     *zx = y2;
 /*<       delta=1./(1.+ta*hr/tr/ha*exp((zx-ppp1)*(1./hr-1./ha))) >*/
-    delta = (float)1. / (*ta * *hr / *tr / *ha * exp((*zx - *ppp1) * ((float)
-	    1. / *hr - (float)1. / *ha)) + (float)1.);
+    delta = 1. / (*ta * *hr / *tr / *ha * exp((*zx - *ppp1) * (1. / *hr - 1. /
+	     *ha)) + 1.);
 /*<       ecart=0 >*/
     ecart = 0.;
 /*<       if(dd.ne.0) ecart=abs((dd-delta)/dd) >*/
@@ -108,7 +107,7 @@ L705:
 	ecart = (d__1 = (*dd - delta) / *dd, abs(d__1));
     }
 /*<       if((ecart.gt.0.75).and.(it.ne.0)) go to 99 >*/
-    if (ecart > (float).75 && *it != 0) {
+    if (ecart > .75 && *it != 0) {
 	goto L99;
     }
 /*<       return >*/
